Implement getting address from Place API
I had to Create and Include API key (developers console to create project and get key and include the key in manifest file and internet permissions); Add dependencies for PlaceAPI to build.gradle(app/build.gradle).
Implemented the initialization of place and required block of code to get the place and then set the editText for the location
I had to set evenListener to call the place autocomplete fragment

FYI i implemented a AES encryption and decryption.