Implemented fireBase write (file upload and Json object)
The overall idea behind the code is that i will first upload the images to firebase storage, download the stringPhotoUrls of the images. The create an instance of our Book pojo, initialize the object and then push to the firebase database

FYI, i had to set the fragmentAdapter for list and favourites in the accountFragment to updated version. As well as setting the DEX file limit for our app in build.gradle