Implemented Sorting the books displayed in the home page based on user location
First i had to reformat the homepage fragment. I implemented a function that set the list of books in the home fragment
I now had to query the database for books from the database at the host main activity then load into the homepage fragment.

The idea behind the algorithm was after getting the user location on app create, i initialize the eventListener that load list of books from the firebase database, after that i then load the fragment.
When you click on the icon, i just load the home fragment with the update list from the valueEventListener