Implemented Displaying the list of books that a user posted in the Listing fragment of the Account page
I had to edit my BooksAdapter to be a Heterogeneous one. Which meas i had to override the getItemViewType to return an integer and based on the integer returnd i will know the particular viewHolder to inflate on the desired recyclerView. So basically ingonre the HomeBooksAdapter. It was renamed BooksAdapter where i did the heterogeneous implementation
I edited the Account fragment to display the userName after the 'Welcome' text

FYI i modified the temp_books_list_item and added a new xml for the Listing Fragment