Implemented favourite and delete menuItem
created xml of heart empty and full, depending on the current drawable of the heart shown, i will show the next state. The delete icon will only show if the person that login is the seller of the book details to show.

FYI I now had to show the book data after i had fetch the pushID of the book details to show.