Implemented loading the favourites books that was selected by the user
The idea here was i created a room database with one table called 'favourites' the stores the pushID of books that is marked favourites by the user.
Also, i had to set the icon to show the state of the book in the detail page whether the book is already marked favourite or not. This ws achieved by room viewmodel to check if the book is in the favourites room database in real time

FYI: room database insertion and delete of book from database was don in background thread, while room database query like loading list of books markd favourites or checking if the book is marked favourites or not was done in ViewModel so that it will be done in real time